+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.21.2] - 2024-04-13
+
+### Fixed
+
+- Health: Report error if version check fails for a required
+  external dependency. This should help with false positives
+  when detecting `rust-analyzer` if the rustup wrapper is installed,
+  but `rust-analyzer` isn't.
+
 ## [4.21.1] - 2024-04-11
 
 ### Fixed